NEW PROJECT FROM 'TRUZZ BLOGG' YOUTUBE CHANNEL LINK: https://www.youtube.com/watch?v=lLMFA7gMeY0&ab_channel=TruzzBlogg
MODIFYED THE ORIGINAL WINDOW NAME TO 'WELCOME TO INBY!'

from 8 to 14 line it's created and well-configurated the window properties and title using LABEL and backgrounds.

then it's a bunch of labels for the field that user will fill out ('username', 'password', 'fullname' and 'age').

finally added the mainloop to avoid quiting the window app.